type fix

+++ MingdaoSDKClass/Model/MDCalendarCategory.h
@@ -13,7 +13,7 @@
 @property (strong, nonatomic) NSString *catID;
 @property (strong, nonatomic) NSString *catName;
 @property (assign, nonatomic) NSInteger displayOrder;
-@property (assign, nonatomic) NSInteger color;
+@property (assign, nonatomic) int color;
 
 - (MDCalendarCategory *)initWithDictionary:(NSDictionary *)aDic;
